Replaced projection matrix to the frustum class
