+++ src/og/renderer/Renderer.js
@@ -91,13 +91,6 @@ const Renderer = function (handler, params) {
     this.renderNodes = {};
 
     /**
-     * Cameras array.
-     * @public
-     * @type {Array.<og.Camera>}
-     */
-    this.cameras = [];
-
-    /**
      * Current active camera.
      * @public
      * @type {og.Camera}
@@ -631,9 +624,9 @@ Renderer.prototype.draw = function () {
 
     var h = this.handler;
 
-    // h.gl.clearColor(0.0, 0.0, 0.0, 1.0);
     h.gl.clearColor(115 / 255, 203 / 255, 249 / 255, 1.0);
-    h.gl.clear(h.gl.COLOR_BUFFER_BIT | h.gl.DEPTH_BUFFER_BIT);
+    h.gl.clear(h.gl.COLOR_BUFFER_BIT);
+    h.gl.clear(h.gl.DEPTH_BUFFER_BIT);
 
     e.dispatch(e.draw, this);
